notification_service/requirements.txt
Flask==2.3.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
twilio==8.10.0
firebase-admin==6.1.0
APScheduler==3.10.1
requests==2.31.0
